test: fold server-side key settings suite
